+++ frontend/techpick/src/components/DndMonitorContext.tsx
@@ -1,0 +1,11 @@
+'use client';
+
+import { PropsWithChildren } from 'react';
+import { usePickToPickDndMonitor, useFolderToFolderDndMonitor } from '@/hooks';
+
+export function DndMonitorContext({ children }: PropsWithChildren) {
+  usePickToPickDndMonitor();
+  useFolderToFolderDndMonitor();
+
+  return <>{children}</>;
+}
